backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
deep-translator==1.11.4
Django==4.2.5
django-summernote==0.8.20.0
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
webencodings==0.5.1
